train/requirements.txt
# tensorflow-datasets==4.0.1
tensorflow-gpu==2.3.1
tensorflow-addons==0.11.2